+++ setup.sh
@@ -427,6 +427,11 @@ elif [[ $sslv = "2" ]]
 then
 setup_letsencrypt
 else
+if which ufw >/dev/null; then
+ufw allow 80/tcp &> /dev/null
+ufw disable &> /dev/null
+ufw --force enable &> /dev/null
+fi
 service nginx restart &>  /dev/null
     print_done "ngninx successfully installed."
 fi
@@ -453,6 +458,7 @@ install_phpmyadmin
 fi
 }
 setup_selfsigned() {
+print_info "Setting up SSL Certificate ..."
 apt-get install openssl &> /dev/null
 mkdir -p /etc/nginx/ssl
 cd /etc/nginx/ssl
@@ -464,7 +470,7 @@ wget -O /etc/nginx/conf.d/default.conf https://raw.githubusercontent.com/eunas/g
 wait
 sed -i "s|        ssl_certificate /etc/letsencrypt/live/domain/fullchain.pem;|        ssl_certificate /etc/nginx/ssl/nginx.crt;|" /etc/nginx/conf.d/default.conf
 sed -i "s|        ssl_certificate_key /etc/letsencrypt/live/domain/privkey.pem;|        ssl_certificate_key /etc/nginx/ssl/nginx.key;|" /etc/nginx/conf.d/default.conf
-sed -i "s|.*server_name.*|        server_name "$d";|" /etc/nginx/conf.d/default.conf
+sed -i "s|        server_name domain www.domain;|        server_name "$d";|" /etc/nginx/conf.d/default.conf
 sed -i "s|        ssl_dhparam /etc/letsencrypt/dhparams.pem;|        ssl_dhparam /etc/nginx/ssl/dhparams.pem;|" /etc/nginx/conf.d/default.conf
 if [ $web = "1" ] ; then
 sed -i "s|.*listen        443.*|       listen        443 ssl;|" /etc/nginx/conf.d/default.conf
@@ -472,6 +478,12 @@ fi
 cd /etc/nginx/ssl/
 openssl dhparam -out dhparams.pem 2048 &> /dev/null
 chmod 600 dhparams.pem
+if which ufw >/dev/null; then
+ufw allow 80/tcp &> /dev/null
+ufw allow 443/tcp &> /dev/null
+ufw disable &> /dev/null
+ufw --force enable &> /dev/null
+fi
 service nginx restart
 print_done "ngninx successfully installed."
 }
@@ -483,14 +495,14 @@ git clone https://github.com/letsencrypt/letsencrypt /opt/letsencrypt &> /dev/nu
 wait
 service nginx stop
 wait
-/opt/letsencrypt/letsencrypt-auto certonly --standalone --agree-tos --email "$mail" -d "$d" -d www."$d" &> /dev/null
+/opt/letsencrypt/letsencrypt-auto certonly --test-cert --standalone --agree-tos --email "$mail" -d "$d" -d www."$d" &> /dev/null
 wait
 rm /etc/nginx/conf.d/default.conf
 wget -O /etc/nginx/conf.d/default.conf https://raw.githubusercontent.com/eunas/gotdeb/master/resources/default-ssl.conf --no-check-certificate &> /dev/null
 wait
 sed -i "s|        ssl_certificate /etc/letsencrypt/live/domain/fullchain.pem;|        ssl_certificate /etc/letsencrypt/live/"$d"/fullchain.pem;|" /etc/nginx/conf.d/default.conf
 sed -i "s|        ssl_certificate_key /etc/letsencrypt/live/domain/privkey.pem;|        ssl_certificate_key /etc/letsencrypt/live/"$d"/privkey.pem;|" /etc/nginx/conf.d/default.conf
-sed -i "s|server_name domain www.domain;|server_name "$d" www."$d";|" /etc/nginx/conf.d/default.conf
+sed -i "s|        server_name domain www.domain;|        server_name "$d" www."$d";|" /etc/nginx/conf.d/default.conf
 if [ $web = "1" ] ; then
 sed -i "s|.*listen        443.*|       listen        443 ssl;|" /etc/nginx/conf.d/default.conf
 fi
@@ -504,6 +516,12 @@ sed -i "s|# webroot-path = /usr/share/nginx/html|webroot-path = /usr/share/nginx
 wget -O /usr/local/sbin/le-renew-webroot https://gist.githubusercontent.com/thisismitch/e1b603165523df66d5cc/raw/fbffbf358e96110d5566f13677d9bd5f4f65794c/le-renew-webroot  &> /dev/null
 chmod +x /usr/local/sbin/le-renew-webroot
 (crontab -l 2>/dev/null; echo "30 2 * * 1 /usr/local/sbin/le-renew-webroot >> /var/log/le-renewal.log") | crontab -
+if which ufw >/dev/null; then
+ufw allow 80/tcp &> /dev/null
+ufw allow 443/tcp &> /dev/null
+ufw disable &> /dev/null
+ufw --force enable &> /dev/null
+fi
 service nginx start
 print_done "ngninx successfully installed."
 }
@@ -714,17 +732,22 @@ echo "yes" > /etc/pure-ftpd/conf/ChrootEveryone
 echo "2" > /etc/pure-ftpd/conf/TLS
 echo "$p" > /etc/pure-ftpd/conf/Bind
 openssl req -x509 -nodes -days 7300 -newkey rsa:2048 -keyout /etc/ssl/private/pure-ftpd.pem -out /etc/ssl/private/pure-ftpd.pem -subj "/C=US/ST=defaultstate/L=defaultcity/O=myorg/CN=localhost"
+if which ufw >/dev/null; then
+ufw allow "$p"/tcp &> /dev/null
+ufw disable &> /dev/null
+ufw --force enable &> /dev/null
+fi
 service pure-ftpd restart &> /dev/null
 print_done "Pure-FTPd with FTPS support successfully installed."
 }
 function install_openvpn {
 if [[ ! -e /dev/net/tun ]]; then
 	print_warn "TUN/TAP is not available"
-	exit
+	exit 1
 fi
 if grep -qs "CentOS release 5" "/etc/redhat-release"; then
 	echo "CentOS 5 is too old and not supported"
-	exit 3
+	exit 2
 fi
 if [[ -e /etc/debian_version ]]; then
 	OS=debian
@@ -736,7 +759,7 @@ elif [[ -e /etc/centos-release || -e /etc/redhat-release ]]; then
 	chmod +x /etc/rc.d/rc.local
 else
 	echo "Looks like you aren't running this installer on a Debian, Ubuntu or CentOS system"
-	exit 4
+	exit 3
 fi
 
 newclient () {
@@ -767,26 +790,26 @@ if [[ -e /etc/openvpn/server.conf ]]; then
 	while :
 	do
 	clear
-		print_info "Looks like OpenVPN is already installed"
-		print_info ""
-		print_info "What do you want to do?"
-		print_info "   1) Add a cert for a new user"
-		print_info "   2) Revoke existing user cert"
-		print_info "   3) Remove OpenVPN"
-		print_info "   4) Exit"
+		echo "Looks like OpenVPN is already installed"
+		echo ""
+		echo "What do you want to do?"
+		echo "   1) Add a cert for a new user"
+		echo "   2) Revoke existing user cert"
+		echo "   3) Remove OpenVPN"
+		echo "   4) Exit"
 		read -p "Select an option [1-4]: " option
 		case $option in
 			1)
-			print_info ""
-			print_info "Tell me a name for the client cert"
-			print_info "Please, use one word only, no special characters"
+			echo ""
+			echo "Tell me a name for the client cert"
+			echo "Please, use one word only, no special characters"
 			read -p "Client name: " -e -i client CLIENT
 			cd /etc/openvpn/easy-rsa/
 			./easyrsa build-client-full $CLIENT nopass
 			# Generates the custom client.ovpn
 			newclient "$CLIENT"
-			print_info ""
-			print_info "Client $CLIENT added, certs available at ~/$CLIENT.ovpn"
+			echo ""
+			echo "Client $CLIENT added, certs available at ~/$CLIENT.ovpn"
 			exit
 			;;
 			2)
@@ -794,12 +817,12 @@ if [[ -e /etc/openvpn/server.conf ]]; then
 			# ...but what can I say, I want some sleep too
 			NUMBEROFCLIENTS=$(tail -n +2 /etc/openvpn/easy-rsa/pki/index.txt | grep -c "^V")
 			if [[ "$NUMBEROFCLIENTS" = '0' ]]; then
-				print_info ""
-				print_info "You have no existing clients!"
+				echo ""
+				echo "You have no existing clients!"
 				exit 5
 			fi
-			print_info ""
-			print_info"Select the existing client certificate you want to revoke"
+			echo ""
+			echo "Select the existing client certificate you want to revoke"
 			tail -n +2 /etc/openvpn/easy-rsa/pki/index.txt | grep "^V" | cut -d '=' -f 2 | nl -s ') '
 			if [[ "$NUMBEROFCLIENTS" = '1' ]]; then
 				read -p "Select one client [1]: " CLIENTNUMBER
@@ -813,6 +836,8 @@ if [[ -e /etc/openvpn/server.conf ]]; then
 			rm -rf pki/reqs/$CLIENT.req
 			rm -rf pki/private/$CLIENT.key
 			rm -rf pki/issued/$CLIENT.crt
+			rm -rf /etc/openvpn/crl.pem
+			cp /etc/openvpn/easy-rsa/pki/crl.pem /etc/openvpn/crl.pem
 			# And restart
 			if pgrep systemd-journal; then
 				systemctl restart openvpn@server.service
@@ -823,12 +848,12 @@ if [[ -e /etc/openvpn/server.conf ]]; then
 					service openvpn restart
 				fi
 			fi
-			print_info ""
-			print_info "Certificate for client $CLIENT revoked"
+			echo ""
+			echo "Certificate for client $CLIENT revoked"
 			exit
 			;;
 			3)
-			print_info ""
+			echo ""
 			read -p "Do you really want to remove OpenVPN? [y/n]: " -e -i n REMOVE
 			if [[ "$REMOVE" = 'y' ]]; then
 				PORT=$(grep '^port ' /etc/openvpn/server.conf | cut -d " " -f 2)
@@ -845,6 +870,11 @@ if [[ -e /etc/openvpn/server.conf ]]; then
 					sed -i "/iptables -I FORWARD -m state --state RELATED,ESTABLISHED -j ACCEPT/d" $RCLOCAL
 				fi
 				sed -i '/iptables -t nat -A POSTROUTING -s 10.8.0.0\/24 -j SNAT --to /d' $RCLOCAL
+				if sestatus | grep "Current mode" | grep -qs "enforcing"; then
+					if [[ "$PORT" != '1194' ]]; then
+						semanage port -d -t openvpn_port_t -p udp $PORT
+					fi
+				fi
 				if [[ "$OS" = 'debian' ]]; then
 					apt-get remove --purge -y openvpn openvpn-blacklist
 				else
@@ -852,11 +882,11 @@ if [[ -e /etc/openvpn/server.conf ]]; then
 				fi
 				rm -rf /etc/openvpn
 				rm -rf /usr/share/doc/openvpn*
-				print_info ""
-				print_info "OpenVPN removed!"
+				echo ""
+				echo "OpenVPN removed!"
 			else
-				print_info ""
-				print_info "Removal aborted!"
+				echo ""
+				echo "Removal aborted!"
 			fi
 			exit
 			;;
@@ -880,11 +910,10 @@ else
 	print_info ""
 	print_info "What DNS do you want to use with the VPN?"
 	print_info "   1) Current system resolvers"
-	print_info "   2) OpenDNS"
-	print_info "   3) Level 3"
+	print_info "   2) Google"
+	print_info "   3) OpenDNS"
 	print_info "   4) NTT"
 	print_info "   5) Hurricane Electric"
-	print_info "   6) Google"
 	read -p "DNS [1-6]: " -e -i 1 DNS
 	print_info ""
 	print_info "Finally, tell me your name for the client cert"
@@ -921,7 +950,7 @@ else
 	./easyrsa build-client-full $CLIENT nopass
 	./easyrsa gen-crl
 	# Move the stuff we need
-	cp pki/ca.crt pki/private/ca.key pki/dh.pem pki/issued/server.crt pki/private/server.key /etc/openvpn
+	cp pki/ca.crt pki/private/ca.key pki/dh.pem pki/issued/server.crt pki/private/server.key /etc/openvpn/easy-rsa/pki/crl.pem /etc/openvpn
 	# Generate server.conf
 	echo "port $PORT
 proto udp
@@ -945,12 +974,12 @@ ifconfig-pool-persist ipp.txt" > /etc/openvpn/server.conf
 		done
 		;;
 		2)
-		echo 'push "dhcp-option DNS 208.67.222.222"' >> /etc/openvpn/server.conf
-		echo 'push "dhcp-option DNS 208.67.220.220"' >> /etc/openvpn/server.conf
+		echo 'push "dhcp-option DNS 8.8.8.8"' >> /etc/openvpn/server.conf
+		echo 'push "dhcp-option DNS 8.8.4.4"' >> /etc/openvpn/server.conf
 		;;
 		3)
-		echo 'push "dhcp-option DNS 4.2.2.2"' >> /etc/openvpn/server.conf
-		echo 'push "dhcp-option DNS 4.2.2.4"' >> /etc/openvpn/server.conf
+		echo 'push "dhcp-option DNS 208.67.222.222"' >> /etc/openvpn/server.conf
+		echo 'push "dhcp-option DNS 208.67.220.220"' >> /etc/openvpn/server.conf
 		;;
 		4)
 		echo 'push "dhcp-option DNS 129.250.35.250"' >> /etc/openvpn/server.conf
@@ -959,10 +988,6 @@ ifconfig-pool-persist ipp.txt" > /etc/openvpn/server.conf
 		5)
 		echo 'push "dhcp-option DNS 74.82.42.42"' >> /etc/openvpn/server.conf
 		;;
-		6)
-		echo 'push "dhcp-option DNS 8.8.8.8"' >> /etc/openvpn/server.conf
-		echo 'push "dhcp-option DNS 8.8.4.4"' >> /etc/openvpn/server.conf
-		;;
 	esac
 	echo "keepalive 10 120
 comp-lzo
@@ -970,7 +995,7 @@ persist-key
 persist-tun
 status openvpn-status.log
 verb 3
-crl-verify /etc/openvpn/easy-rsa/pki/crl.pem" >> /etc/openvpn/server.conf
+crl-verify crl.pem" >> /etc/openvpn/server.conf
 	# Enable net.ipv4.ip_forward for the system
 	if [[ "$OS" = 'debian' ]]; then
 		sed -i 's|#net.ipv4.ip_forward=1|net.ipv4.ip_forward=1|' /etc/sysctl.conf
@@ -1006,6 +1031,16 @@ crl-verify /etc/openvpn/easy-rsa/pki/crl.pem" >> /etc/openvpn/server.conf
 		sed -i "1 a\iptables -I INPUT -p udp --dport $PORT -j ACCEPT" $RCLOCAL
 		sed -i "1 a\iptables -I FORWARD -s 10.8.0.0/24 -j ACCEPT" $RCLOCAL
 		sed -i "1 a\iptables -I FORWARD -m state --state RELATED,ESTABLISHED -j ACCEPT" $RCLOCAL
+	fi
+	# If SELinux is enabled and a custom port was selected, we need this
+	if sestatus | grep "Current mode" | grep -qs "enforcing"; then
+		if [[ "$PORT" != '1194' ]]; then
+			# semanage isn't available in CentOS 6 by default
+			if ! which semanage > /dev/null 2>&1; then
+				yum install policycoreutils-python -y
+			fi
+			semanage port -a -t openvpn_port_t -p udp $PORT
+		fi
 	fi
 	# And finally, restart OpenVPN
 	if [[ "$OS" = 'debian' ]]; then
@@ -1052,6 +1087,11 @@ remote-cert-tls server
 comp-lzo
 verb 3" > /etc/openvpn/client-common.txt
 	# Generates the custom client.ovpn
+if which ufw >/dev/null; then
+ufw allow "$PORT"/tcp &> /dev/null
+ufw disable &> /dev/null
+ufw --force enable &> /dev/null
+fi
 	newclient "$CLIENT"
 	print_info ""
 	print_info "Finished!"
@@ -1175,7 +1215,11 @@ if [ $(plain_version) = "8" ]; then
 sed -i "s|.*auth_param basic program.*|auth_param basic program /usr/lib/squid3/basic_ncsa_auth /etc/squid3/squid_passwd|" /etc/squid3/squid.conf
 fi
 htpasswd -b -c /etc/squid3/squid_passwd $u $p
-
+if which ufw >/dev/null; then
+ufw allow "$p"/tcp &> /dev/null
+ufw disable &> /dev/null
+ufw --force enable &> /dev/null
+fi
 service squid3 restart
 
 clear
@@ -1494,7 +1538,7 @@ print_info "https://gotdeb.com"
 print_info ""
 print_info "Credits: Xeoncross, mikel, Falko Timme, road warrior, Nyr and many others",
 print_info ""
-print_info "Version 1.6.3"
+print_info "Version 1.6.4"
 }
 function system_tests {
 	print_info "Classic I/O test"
@@ -1505,23 +1549,20 @@ function system_tests {
 	print_info "wget cachefly.cachefly.net/100mb.test -O 100mb.test && rm -fr 100mb.test"
 	wget cachefly.cachefly.net/100mb.test -O 100mb.test && rm -fr 100mb.test
 }
-function configure_aria2 {
-#########################
-check_install nginx 1 "nginx is already installed. Please remove it before installing Aria2."
-php=n
-db=n
-db1=n
-install_webserver
-#########################
+configure_aria2() {
 if which aria2c >/dev/null; then
 print_warn "Aria2 is already installed."
 exit 1
 fi
-if [ -x /usr/sbin/nginx ]; then
-if [[ ! -f /usr/sbin/php5-fpm ]] && [[ ! -f /usr/sbin/php-fpm7.0 ]] && [[ ! -f /usr/bin/hhvm ]]; then
- print_warn "PHP or HHVM is not installed."
-exit 1
-fi
+check_install nginx 1 "nginx is already installed. Please remove it before installing Aria2."
+print_info "Enter a secret token"
+read -s secret
+print_info "Enter a port for the Aria2 Daemon"
+read aport
+php=n
+db=n
+db1=n
+install_webserver
 print_info "Installing Aria2 (This might take some time, please be patient...)"
 file="/etc/apt/sources.list.d/debian-testing.list"
 if [ ! -f "$file" ]
@@ -1554,7 +1595,7 @@ summary-interval=120
 daemon=true
 enable-rpc=true
 enable-dht=true
-rpc-listen-port=6800
+rpc-listen-port=
 rpc-listen-all=true
 max-concurrent-downloads=3
 http-auth-challenge=true
@@ -1566,9 +1607,7 @@ timeout=600
 retry-wait=30
 max-tries=50
 EOM
-print_info "Enter a secret token"
-read -s secret
-print_info "Configuring Aria2..."
+sed -i "s|rpc-listen-port=|rpc-listen-port=$aport|" /usr/share/aria2/aria2.conf
 touch /etc/init.d/aria2
 /bin/cat <<"EOM" >/etc/init.d/aria2
 #! /bin/sh
@@ -1610,19 +1649,46 @@ RETVAL=1
 esac
 exit $RETVAL
 EOM
-sed -i "s|.*aria2c --daemon=true --enable-rpc --rpc-listen-all --rpc-secret=secret.*|aria2c --daemon=true --enable-rpc --rpc-listen-all --rpc-secret=$secret -D --conf-path=/usr/share/aria2/aria2.conf;|" /etc/init.d/aria2
+sed -i "s|.*aria2c --daemon=true --enable-rpc --rpc-listen-all --rpc-secret=secret.*|aria2c --daemon=true --enable-rpc --rpc-listen-all --rpc-secret=$secret -D --conf-path=/usr/share/aria2/aria2.conf|" /etc/init.d/aria2
 chmod +x /etc/init.d/aria2
 update-rc.d aria2 defaults &> /dev/null
 git clone https://github.com/ziahamza/webui-aria2.git /usr/share/nginx/html/aria2 &> /dev/null
+wait
+if which ufw >/dev/null; then
+ufw allow "$aport"/tcp &> /dev/null
+ufw allow 80/tcp &> /dev/null
+ufw disable &> /dev/null
+ufw --force enable &> /dev/null
+fi
 service aria2 start &> /dev/null
 wait
 rm -rf /tmp/aria2
-print_done "Aria2 has been installed"
-print_done "Access it at http://$(get_external_ip)/aria2"
-print_done "Your secret token is $secret"
-else
-print_warn "No webserver installed. Aborting"
+if [[ $ssl = "y" ]]
+then
+if [[ $sslv = "1" ]]
+then
+sed -i "s|.*aria2c --daemon=true --enable-rpc --rpc-listen-all --rpc-secret.*|aria2c --daemon=true --enable-rpc --rpc-certificate=/etc/nginx/ssl/nginx.crt --rpc-private-key=/etc/nginx/ssl/nginx.key --rpc-secure=true --rpc-listen-all --rpc-secret=$secret -D --conf-path=/usr/share/aria2/aria2.conf|" /etc/init.d/aria2
+elif [[ $sslv = "2" ]]
+then
+sed -i "s|.*aria2c --daemon=true --enable-rpc --rpc-listen-all --rpc-secret.*|aria2c --daemon=true --enable-rpc --rpc-certificate=/etc/letsencrypt/live/"$d"/fullchain.pem --rpc-private-key=/etc/letsencrypt/live/"$d"/privkey.pem --rpc-secure=true --rpc-listen-all --rpc-secret=$secret -D --conf-path=/usr/share/aria2/aria2.conf|" /etc/init.d/aria2
 fi
+if which ufw >/dev/null; then
+ufw allow "$aport"/tcp &> /dev/null
+ufw allow 80/tcp &> /dev/null
+ufw allow 443/tcp &> /dev/null
+ufw disable &> /dev/null
+ufw --force enable &> /dev/null
+fi
+sed -i "s|.*port: uri.port(),|      port: $aport,|" /usr/share/nginx/html/aria2/js/services/rpc/rpc.js
+sed -i "s|.*port: 6800,|      port: $aport,|" /usr/share/nginx/html/aria2/js/services/rpc/rpc.js
+systemctl daemon-reload
+service aria2 stop
+sleep 3
+service aria2 start
+fi
+print_done "Aria2 has been installed"
+print_done "Access it at http://$d/aria2"
+print_done "Your secret token is $secret"
 }
 function get_linuxdash {
 check_install nginx 0 "Please install nginx first."
@@ -1647,7 +1713,7 @@ else
 python /home/speedtest-cli  --share
 fi
 }
-function install_softether {
+install_softether() {
 check_install softether 1 "SoftEtherVPN is already installed" v
 print_info "Running pre checks, this might take a while..."
 apt-get update &> /dev/null
@@ -1823,6 +1889,11 @@ echo iptables-persistent iptables-persistent/autosave_v4 boolean true | debconf-
 echo iptables-persistent iptables-persistent/autosave_v6 boolean true | debconf-set-selections
 apt-get install iptables-persistent -y &> /dev/null
 sed -i "s|.*#user=.*|user=root|" /etc/dnsmasq.conf
+if which ufw >/dev/null; then
+ufw allow "$port"/tcp &> /dev/null
+ufw disable &> /dev/null
+ufw --force enable &> /dev/null
+fi
 print_info "Restarting services..."
 /etc/init.d/vpnserver restart &> /dev/null
 /etc/init.d/dnsmasq restart &> /dev/null
@@ -1835,7 +1906,7 @@ print_done "SoftEtherVPN has been installed"
 print_done "Please see the wiki https://github.com/eunas/gotdeb/wiki/SoftEtherVPN"
 print_done "For further information."
 }
-function install_remotedesktop {
+install_remotedesktop() {
 check_install x2goserver 1 "X2Go Server is already installed." v
 print_info "Install X2GO. Please wait ..."
 apt-key adv --recv-keys --keyserver keys.gnupg.net E1F958385BFE2B6E
@@ -1863,40 +1934,48 @@ print_done "X2Go client can be downloaded from"
 print_done "http://wiki.x2go.org/doku.php/download:start"
 }
 secure_system() {
-check_install fail2ban 1 "fail2ban is already installed."
+install_fail2ban() {
 while true; do
+check_install fail2ban 1 "fail2ban is already installed." v
 print_info "This will install fail2ban, change the ssh port,"
 print_info "permit ssh root login and create a new user"
 print_info "Are you sure you want to continue ? [y/n]"
 read -n1 choice
 case $choice in
 y|Y|yes|Yes|YES)
-print_info "Installing fail2ban...."
+print_info "Name for the new user:"
+read u
+mkdir -p "/home/$u"
+chmod 750 /home/$u
+useradd -d /home/$u $u
+chown -R $u /home/$u
+wait
+passwd $u
+print_info "Choose a new ssh port (Press enter to skip)"
+read p
+print_done "User $u added with home dir /home/$u"
+print_info "Installing..."
 apt-get update &> /dev/null
 apt-get install fail2ban -y &> /dev/null
 wait
 cp /etc/fail2ban/jail.conf /etc/fail2ban/jail.local
 sed -i "s|.*PermitRootLogin yes.*|PermitRootLogin no|"  /etc/ssh/sshd_config
-print_info "Name for the new user:"
-read u
-useradd -d /home/$username $u
-wait
-mkdir -p "/home/$u"
-chmod 750 /home/$u
-chown -R $u /home/$u
-wait
-passwd $u
-print_done "User $username added with home dir /home/$u"
-print_info "Choose a new ssh port (Press enter to skip)"
-read p
 if [[ -n "$p" ]] ; then
 sed -i "s|.*Port.*|Port $p|"  /etc/ssh/sshd_config
+elif [[ -z "$p" ]] ; then
+PORT=$(cat /etc/ssh/sshd_config | grep Port)
+unset p
+p=${PORT#*Port }
+fi
+if which ufw >/dev/null; then
+ufw allow $p &> /dev/null
+ufw disable &> /dev/null
+ufw --force enable &> /dev/null
 fi
 print_info "Restarting services...."
 service fail2ban restart &> /dev/null
 wait
 service ssh restart &> /dev/null
-wait
 print_done "Install complete."
 print_done "Please check that your new user can login with ssh before closing this session."
 break
@@ -1906,6 +1985,72 @@ break
 ;;
      *)
      echo "That is not a valid choice."
+     ;;
+esac
+done
+}
+install_ufw() {
+check_install ufw 1 "ufw  is already installed." v
+print_info "Ufw will disable all incoming ports except for ssh"
+print_info "Ports can be opened using ufw allow port-number"
+sleep 3
+print_info "Installing..."
+apt-get update &> /dev/null
+apt-get install ufw -y &> /dev/null
+PORT=$(cat /etc/ssh/sshd_config | grep Port)
+unset p
+p=${PORT#*Port }
+if [[ ! /proc/net/if_inet6 ]]
+ then
+ipv6=0
+else
+ipv6=1
+fi
+if [[ $ipv6 = "1" ]]
+then
+sed -i "s|IPV6=.*|IPV6=yes|" /etc/default/ufw
+fi
+ufw default deny incoming &> /dev/null
+ufw default allow outgoing &> /dev/null
+ufw allow "$p" &> /dev/null
+ufw --force enable &> /dev/null
+print_done "Install complete."
+}
+unattended_upgrades() {
+print_info "Setting up Unattended Upgrades..."
+if [ -f /etc/cron.daily/apt.disabled ]; then
+    mv /etc/cron.daily/apt.disabled /etc/cron.daily/apt
+fi
+apt-get upgrade &> /dev/null
+echo unattended-upgrades unattended-upgrades/enable_auto_updates boolean true | debconf-set-selections &> /dev/null
+apt-get install unattended-upgrades -y &> /dev/null
+print_done "Your installation is now configured to automaticly install critical updates."
+}
+while true; do
+print_info "Select an option"
+print_info "1) Install fail2ban"
+print_info "2) Install UFW"
+print_info "3) Unattended Upgrades"
+print_info "e) Exit"
+read -n1 choice
+case $choice in
+1)
+install_fail2ban
+break
+;;
+2)
+install_ufw
+break
+;;
+3)
+unattended_upgrades
+break
+;;
+e|E)
+break
+;;
+     *)
+     print_warn "That is not a valid choice, try a number from 1 to 3."
      ;;
 esac
 done
@@ -1936,6 +2081,13 @@ wget --no-check-certificate -O /etc/init.d/plexmediaserver https://raw.githubuse
 wait
 chmod +x /etc/init.d/plexmediaserver
 update-rc.d plexmediaserver defaults &> /dev/null
+if which ufw >/dev/null; then
+ufw allow 32400/tcp &> /dev/null
+ufw disable &> /dev/null
+ufw --force enable &> /dev/null
+fi
+service plexmediaserver restart &> /dev/null
+wait
 print_done "Plex media server has been installed. You can access it at http://$(get_ip):32400/web To get access to the server settings please setup a VPN on the server and access it's localip using that."
 }
 function setup_observium {
@@ -1964,19 +2116,22 @@ esac
 done
 }
 function install_observium_server {
-if [ -x /usr/sbin/nginx ]; then
-if [[ ! -f /usr/sbin/php5-fpm ]] && [[ ! -f /usr/sbin/php-fpm7.0 ]] && [[ ! -f /usr/bin/hhvm ]]; then
- print_warn "PHP or HHVM is not installed."
-exit 1
+print_info "Select a database server"
+print_info "1) MariaDB"
+print_info "2) MySQL"
+read -n 1 dbs
+if [[ $dbs = "1" ]] ; then
+db=y
+db1=n
+elif [[ $dbs = "2" ]] ; then
+db=n
+db1=y
 fi
-if ((! $(ps -ef | grep -v grep | grep mysql | wc -l) > 0 ))
-then
-        print_warn "The MySQL server is stopped or not installed.";
-        exit 1
-
-fi
+php=y
+install_webserver
+wait
 rand=$(cat /dev/urandom | tr -dc 'a-zA-Z0-9' | fold -w 4 | head -n 1)
-u=observ_$rand
+u=observium
 p=$(cat /dev/urandom | tr -dc 'a-zA-Z0-9' | fold -w 12 | head -n 1)
 #EXPECTED_ARGS=3
 #E_BADARGS=65
@@ -1986,12 +2141,44 @@ Q2="GRANT USAGE ON *.* TO $u@localhost IDENTIFIED BY '$p';"
 Q3="GRANT ALL PRIVILEGES ON observium.* TO $u@localhost;"
 Q4="FLUSH PRIVILEGES;"
 SQL="${Q1}${Q2}${Q3}${Q4}"
-print_info "Enter mysql root password"
-$MYSQL -uroot -p -e "$SQL"
 print_info "Installing observium..."
-DEBIAN_FRONTEND=noninteractive apt-get upgrade &> /dev/null
-DEBIAN_FRONTEND=noninteractive apt-get install -y php5-snmp php-pear snmp graphviz php5-json rrdtool fping imagemagick whois mtr-tiny nmap ipmitool python-mysqldb curl &> /dev/null
+$MYSQL -uroot -p$dbpass -e "$SQL"
+apt-get update &> /dev/null
+rm /etc/nginx/conf.d/default.conf
+wget -O /etc/nginx/conf.d/default.conf https://raw.githubusercontent.com/eunas/gotdeb/master/resources/default.conf --no-check-certificate &> /dev/null
+if [[ -e /usr/sbin/php-fpm7.0 ]]; then
+apt-get install php7.0-snmp php-pear snmp graphviz php7.0-json rrdtool fping imagemagick whois mtr-tiny nmap ipmitool python-mysqldb curl -y &> /dev/null
+elif [[ -e /usr/sbin/php5-fpm ]] || [[ -e /usr/bin/hhvm ]] ; then
+apt-get install php5-snmp php-pear snmp graphviz php5-json rrdtool fping imagemagick whois mtr-tiny nmap ipmitool python-mysqldb curl -y &> /dev/null
+fi
+if [[ $ssl = "y" ]]; then
+rm /etc/nginx/conf.d/default.conf
+wget -O /etc/nginx/conf.d/default.conf https://raw.githubusercontent.com/eunas/gotdeb/master/resources/observium-ssl.conf --no-check-certificate &> /dev/null
 wait
+if [ $sslv = "1" ] ; then
+sed -i "s|        ssl_certificate /cert.crt;|        ssl_certificate /etc/nginx/ssl/nginx.crt;|" /etc/nginx/conf.d/default.conf
+sed -i "s|        ssl_certificate_key /privkey.key;|        ssl_certificate_key /etc/nginx/ssl/nginx.key;|" /etc/nginx/conf.d/default.conf
+sed -i "s|        ssl_dhparam /dhparams.pem;|        ssl_dhparam /etc/nginx/ssl/dhparams.pem;|" /etc/nginx/conf.d/default.conf
+fi
+if [ $sslv = "2" ] ; then
+sed -i "s|        ssl_certificate /cert.crt;|        ssl_certificate /etc/letsencrypt/live/"$d"/fullchain.pem;|" /etc/nginx/conf.d/default.conf
+sed -i "s|        ssl_certificate_key /privkey.key;|        ssl_certificate_key /etc/letsencrypt/live/"$d"/privkey.pem;|" /etc/nginx/conf.d/default.conf
+sed -i "s|        ssl_dhparam /dhparams.pem;|        ssl_dhparam /etc/letsencrypt/dhparams.pem;|" /etc/nginx/conf.d/default.conf
+fi
+if [ $web = "1" ] ; then
+sed -i "s|       listen        443 ssl http2;|       listen        443 ssl;|" /etc/nginx/conf.d/default.conf
+fi
+fi
+if [[ $phpv = "1" ]] ; then
+sed -i "s|.*fastcgi_pass unix:/var/run/php.sock;|        fastcgi_pass unix:/var/run/php5-fpm.sock;|" /etc/nginx/conf.d/default.conf
+elif [[ $phpv = "2" ]] ; then
+sed -i "s|.*fastcgi_pass unix:/var/run/php.sock;|        fastcgi_pass unix:/var/run/php/php7.0-fpm.sock;|" /etc/nginx/conf.d/default.conf
+elif [[ $phpv = "3" ]] ; then
+sed -i "s|.*fastcgi_pass unix:/var/run/php.sock;|        fastcgi_pass unix:/var/run/hhvm/hhvm.sock;|" /etc/nginx/conf.d/default.conf
+fi
+sed -i "s|        server_name domain www.domain;|        server_name "$d";|" /etc/nginx/conf.d/default.conf
+sed -i "5s|.*root /usr/share/nginx/html;|        root /opt/observium/html;|" /etc/nginx/conf.d/default.conf
+sed -i "16s|.*root /usr/share/nginx/html;|        root /opt/observium/html;|" /etc/nginx/conf.d/default.conf
 mkdir -p /opt/observium && cd /opt
 wget -P /opt/ http://www.observium.org/observium-community-latest.tar.gz &> /dev/null
 tar zxvf /opt/observium-community-latest.tar.gz -C /opt &> /dev/null
@@ -1999,22 +2186,19 @@ cp /opt/observium/config.php.default /opt/observium/config.php
 sed -i "s|USERNAME|"$u"|" /opt/observium/config.php
 sed -i "s|PASSWORD|"$p"|" /opt/observium/config.php
 mkdir -p /opt/observium/rrd
+mkdir /opt/observium/logs
 chown www-data:www-data /opt/observium/rrd
 cd /opt/observium
-php includes/update/update.php &> /dev/null
-rm /etc/nginx/conf.d/default.conf
-wget -P /etc/nginx/conf.d/ https://raw.githubusercontent.com/eunas/gotdeb/master/resources/default.conf &> /dev/null
-sed -i "s|server_name _;|server_name "$(get_ip)";|" /etc/nginx/conf.d/default.conf
-service nginx restart &> /dev/null
-service php5-fpm restart &> /dev/null
 randp=$(cat /dev/urandom | tr -dc 'a-zA-Z0-9' | fold -w 12 | head -n 1)
-php adduser.php admin "$randp" 10
+php /opt/observium/discovery.php -u &> /dev/null
+php adduser.php admin "$randp" 10 &> /dev/null
 touch /etc/cron.d/observium
+service nginx restart &> /dev/null
 echo "33  */6   * * *   root    /opt/observium/discovery.php -h all >> /dev/null 2>&1" >> /etc/cron.d/observium
 echo "*/5 *     * * *   root    /opt/observium/discovery.php -h new >> /dev/null 2>&1" >> /etc/cron.d/observium
 echo "*/5 *     * * *   root    /opt/observium/poller-wrapper.py 2 >> /dev/null 2>&1" >> /etc/cron.d/observium
 print_done "---------------------------------------------------------------"
-print_done "Observium has been installed. Login at http://$(get_external_ip)"
+print_done "Observium has been installed. Login at http://$d"
 print_done "Login details:"
 print_done "Username: admin"
 print_done "Password: $randp"
@@ -2029,10 +2213,7 @@ print_done "---------------------------------------------------------------"
 print_done ""
 print_done ""
 print_done ""
-else
-print_warn "Observium requires a webserver, PHP and a database server. Aborting."
 exit 1
-fi
 }
 function install_observium_client {
 #!/bin/bash
@@ -2072,6 +2253,11 @@ END
 wget -O distro https://raw.githubusercontent.com/eunas/gotdeb/master/resources/observium_distro --no-check-certificate &> /dev/null
 mv distro /usr/bin/distro
 chmod +x /usr/bin/distro
+if which ufw >/dev/null; then
+ufw allow "$port"/tcp &> /dev/null
+ufw disable &> /dev/null
+ufw --force enable &> /dev/null
+fi
 /etc/init.d/snmpd restart &> /dev/null
 print_done "#########################################################"
 print_done "##           !! !! Installation Complete !! !!         ##"
@@ -2119,6 +2305,11 @@ sed -i "s|.*rpc-port.*|    \"rpc-port\": $trp,|" /etc/transmission-daemon/settin
 sed -i "s|.*rpc-bind-address.*|    \"rpc-bind-address\": \"$(get_ip)\",|" /etc/transmission-daemon/settings.json
 sed -i "s|.*rpc-username.*|    \"rpc-username\": \"$transu\",|" /etc/transmission-daemon/settings.json
 sed -i "s|\"rpc-whitelist\": \"127.0.0.1\",|\"rpc-whitelist\": \"127.0.0.1,*.*.*.*\",|" /etc/transmission-daemon/settings.json
+if which ufw >/dev/null; then
+ufw allow "$trp"/tcp &> /dev/null
+ufw disable &> /dev/null
+ufw --force enable &> /dev/null
+fi
 service transmission-daemon start &> /dev/null
 wait
 clear
@@ -2225,30 +2416,23 @@ cp config.example.js config.js
 wget -O /etc/nginx/conf.d/ghost.conf https://raw.githubusercontent.com/eunas/gotdeb/master/resources/ghost.conf --no-check-certificate  &> /dev/null
 wait
 mv /etc/nginx/conf.d/default.conf /etc/nginx/conf.d/default-backup
+if [[  ssl = "y" ]]; then
 if [ $sslv = "1" ] ; then
-rm /etc/nginx/conf.d/ghost.conf
-wget -O /etc/nginx/conf.d/ghost.conf https://raw.githubusercontent.com/eunas/gotdeb/master/resources/ghost-ssl.conf --no-check-certificate  &> /dev/null
-wait
 sed -i "s|    ssl_dhparam|    ssl_dhparam /etc/nginx/ssl/dhparams.pem;|" /etc/nginx/conf.d/ghost.conf
+fi
+if [ $sslv = "2" ] ; then
+sed -i "s|    ssl_certificate /etc/nginx/ssl/nginx.crt;|    ssl_certificate /etc/letsencrypt/live/"$d"/fullchain.pem;|" /etc/nginx/conf.d/ghost.conf
+sed -i "s|    ssl_certificate_key /etc/nginx/ssl/nginx.key;|    ssl_certificate_key /etc/letsencrypt/live/"$d"/privkey.pem;|" /etc/nginx/conf.d/ghost.conf
+sed -i "s|    ssl_dhparam|    ssl_dhparam /etc/letsencrypt/dhparams.pem;|" /etc/nginx/conf.d/ghost.conf
+fi
+if [ $web = "1" ] ; then
+sed -i "s|    listen 443 ssl http2;|    listen 443 ssl;|" /etc/nginx/conf.d/ghost.conf
+fi
+fi
 if [[ $phpv = "2" ]] ; then
 sed -i "s|.*fastcgi_pass unix:/var/run/php5-fpm.sock;|fastcgi_pass unix:/var/run/php/php7.0-fpm.sock;|" /etc/nginx/conf.d/ghost.conf
 elif [[ $phpv = "3" ]] ; then
 sed -i "s|        fastcgi_pass unix:/var/run/php5-fpm.sock;|        fastcgi_pass unix:/var/run/hhvm/hhvm.sock;|" /etc/nginx/conf.d/ghost.conf
-fi
-if [ $web = "1" ] ; then
-sed -i "s|    listen 443 ssl http2;|    listen 443 ssl;|" /etc/nginx/conf.d/ghost.conf
-fi
-fi
-if [ $sslv = "2" ] ; then
-rm /etc/nginx/conf.d/ghost.conf
-wget -O /etc/nginx/conf.d/ghost.conf https://raw.githubusercontent.com/eunas/gotdeb/master/resources/ghost-ssl.conf --no-check-certificate  &> /dev/null
-wait
-sed -i "s|    ssl_certificate /etc/nginx/ssl/nginx.crt;|    ssl_certificate /etc/letsencrypt/live/"$d"/fullchain.pem;|" /etc/nginx/conf.d/ghost.conf
-sed -i "s|    ssl_certificate_key /etc/nginx/ssl/nginx.key;|    ssl_certificate_key /etc/letsencrypt/live/"$d"/privkey.pem;|" /etc/nginx/conf.d/ghost.conf
-sed -i "s|    ssl_dhparam|    ssl_dhparam /etc/letsencrypt/dhparams.pem;|" /etc/nginx/conf.d/ghost.conf
-if [ $web = "1" ] ; then
-sed -i "s|    listen 443 ssl http2;|    listen 443 ssl;|" /etc/nginx/conf.d/ghost.conf
-fi
 fi
 if [[ $dbs = "2" ]] || [[ $dbs = "3" ]] ; then
 rand=$(cat /dev/urandom | tr -dc 'a-zA-Z0-9' | fold -w 4 | head -n 1)